+++ katdal/h5datav1.py
@@ -1,5 +1,5 @@
 ################################################################################
-# Copyright (c) 2011-2022, National Research Foundation (SARAO)
+# Copyright (c) 2011-2023, National Research Foundation (SARAO)
 #
 # Licensed under the BSD 3-Clause License (the "License"); you may not use
 # this file except in compliance with the License. You may obtain a copy
@@ -131,7 +131,7 @@ class H5DataV1(DataSet):
         self.dump_period = 1.0 / corr_group.attrs['dump_rate_hz']
         self._segments = np.cumsum([0] + [len(s['timestamps']) for s in self._scan_groups])
         num_dumps = self._segments[-1]
-        self._time_keep = np.ones(num_dumps, dtype=np.bool)
+        self._time_keep = np.ones(num_dumps, dtype=bool)
         data_timestamps = self.timestamps
         if data_timestamps[0] < 1e9:
             logger.warning("File '%s' has invalid first correlator timestamp (%f)", filename, data_timestamps[0])
@@ -421,7 +421,7 @@ class H5DataV1(DataSet):
         """
         # Tell the user that there are no flags in the h5 file
         logger.warning("No flags in HDF5 v1 data files, returning array of zero flags")
-        falses = LazyTransform('falses', lambda data, keep: np.zeros_like(data, dtype=np.bool), dtype=np.bool)
+        falses = LazyTransform('falses', lambda data, keep: np.zeros_like(data, dtype=bool), dtype=bool)
         return ConcatenatedLazyIndexer(self._vis_indexers(), transforms=[falses])
 
     @property