+++ NEWS.rst
@@ -1,6 +1,20 @@
 History
 =======
 
+0.22 (2023-11-28)
+-----------------
+* Restore np.bool in Numba averaging function to prevent mvftoms crash (#370)
+* Replace underscores with dashes when loading old buckets from RDBs (#370)
+* Select multiple targets with same name to avoid dropped scans in MS (#369)
+* Support on-the-fly (OTF) scans in mvftoms (#366)
+
+0.21 (2023-05-12)
+-----------------
+* Fix support for numpy >= 1.24 and move unit tests from nose to pytest (#361)
+* Complete rewrite of S3ChunkStore retries for more robust archive downloads (#363)
+* Remove IMAGING_WEIGHT column full of zeroes from MS (#356)
+* Improve tests with ES256-encoded JWT tokens and more robust MinIO health check (#360)
+
 0.20.1 (2022-04-29)
 -------------------
 * Fix broken `dataset.vis[n]` due to DaskLazyIndexer / ChunkStore interaction (#355)